# apidog.openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Dragon Lap MVP Backend API",
    "description": "Single-tenant laptop store backend (NestJS + Prisma + JWT).",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Local development"
    }
  ],
...
  "security": [
    {
      "cookieAuth": []
    }
  ],
  "paths": {
    "/auth/register": {
      "post": {
        "tags": ["Auth"],
        "summary": "Register customer account",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RegisterDto" },
              "example": {
                "fullName": "Ahmed Ali",
                "email": "ahmed@example.com",
                "password": "secret123",
                "phone": "01011223344"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Registered — tokens are set as HTTP-only cookies (access_token, refresh_token)",
            "headers": {
              "Set-Cookie": {
                "description": "access_token (5 min) and refresh_token (7 days) HTTP-only cookies",
                "schema": { "type": "string" }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "user": { "$ref": "#/components/schemas/UserPublic" }
                  }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "tags": ["Auth"],
        "summary": "Logout — revokes refresh token and clears auth cookies",
        "responses": {
          "201": {
            "description": "Logged out",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": { "type": "boolean" }
                  }
                },
                "example": { "ok": true }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": ["Auth"],
        "summary": "Login customer or admin",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/LoginDto" },
              "example": {
                "email": "admin@dragonlap.local",
                "password": "admin123456"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Logged in — tokens are set as HTTP-only cookies (access_token, refresh_token)",
            "headers": {
              "Set-Cookie": {
                "description": "access_token (5 min) and refresh_token (7 days) HTTP-only cookies",
                "schema": { "type": "string" }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "user": { "$ref": "#/components/schemas/UserLoginView" }
                  }
                },
                "example": {
                  "user": {
                    "id": "4c3fdd44-f9de-44b3-b3fc-a1818eeff511",
                    "fullName": "Store Admin",
                    "email": "admin@dragonlap.local",
                    "role": "ADMIN"
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
...
    "/admin/categories": {
      "get": {
        "tags": ["Admin", "Categories"],
        "summary": "List categories (admin)",
        "responses": {
          "200": {
            "description": "Categories",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Category" }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      },
      "post": {
        "tags": ["Admin", "Categories"],
        "summary": "Create category (admin)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateCategoryDto" },
              "example": {
                "name": "Gaming Laptops",
                "slug": "gaming-laptops",
                "description": "High-performance laptops for gaming and graphics workloads"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Category" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      }
    },
    "/admin/categories/{id}": {
      "get": {
        "tags": ["Admin", "Categories"],
        "summary": "Get category by ID (admin)",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": {
            "description": "Category",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Category" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "patch": {
        "tags": ["Admin", "Categories"],
        "summary": "Update category (admin)",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateCategoryDto" },
              "example": {
                "name": "Premium Ultrabooks",
                "description": "Lightweight premium laptops"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Category" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "delete": {
        "tags": ["Admin", "Categories"],
        "summary": "Delete category (admin)",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": {
            "description": "Deleted",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string" }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
...
    "/cart": {
      "get": {
        "tags": ["Cart"],
        "summary": "Get current customer cart",
        "responses": {
          "200": {
            "description": "Cart items",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/CartItemWithLaptop" }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      }
    },
    "/cart/items": {
      "post": {
        "tags": ["Cart"],
        "summary": "Add item to cart",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/AddCartItemDto" },
              "example": {
                "laptopId": "78e13f7e-c53f-4f03-95f1-3f5d2fadb6c2",
                "quantity": 2
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Added to cart",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CartItemWithLaptop" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/cart/items/{id}": {
      "patch": {
        "tags": ["Cart"],
        "summary": "Update cart item quantity",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateCartItemDto" },
              "example": { "quantity": 1 }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated cart item",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CartItemWithLaptop" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "delete": {
        "tags": ["Cart"],
        "summary": "Remove cart item",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": {
            "description": "Removed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string" }
                  },
                  "example": {
                    "message": "Item removed from cart"
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/cart/guest": {
      "get": {
        "tags": ["Cart"],
        "summary": "Get guest cart using guest session cookie",
        "security": [],
        "responses": {
          "200": {
            "description": "Guest cart items",
            "headers": {
              "Set-Cookie": {
                "description": "guest_session_id HTTP-only cookie is set if missing",
                "schema": { "type": "string" }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/CartItemWithLaptop" }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/cart/guest/items": {
      "post": {
        "tags": ["Cart"],
        "summary": "Add item to guest cart",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/AddCartItemDto" },
              "example": {
                "laptopId": "78e13f7e-c53f-4f03-95f1-3f5d2fadb6c2",
                "quantity": 1
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Added to guest cart",
            "headers": {
              "Set-Cookie": {
                "description": "guest_session_id HTTP-only cookie is set if missing",
                "schema": { "type": "string" }
              }
            },
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CartItemWithLaptop" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/cart/guest/items/{id}": {
      "patch": {
        "tags": ["Cart"],
        "summary": "Update guest cart item quantity",
        "security": [],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateCartItemDto" },
              "example": { "quantity": 2 }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated guest cart item",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CartItemWithLaptop" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "delete": {
        "tags": ["Cart"],
        "summary": "Remove guest cart item",
        "security": [],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": {
            "description": "Removed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string" }
                  },
                  "example": {
                    "message": "Item removed from cart"
                  }
                }
              }
            }
          },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "access_token",
        "description": "HTTP-only JWT cookie set automatically on login/register. Send requests with credentials (withCredentials: true)."
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Validation or business rule failure",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" },
            "example": {
              "statusCode": 400,
              "message": "Requested quantity exceeds stock",
              "error": "Bad Request"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Missing or invalid token",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" },
            "example": {
              "statusCode": 401,
              "message": "Unauthorized",
              "error": "Unauthorized"
            }
          }
        }
      },
      "Forbidden": {
        "description": "Role/ownership denied",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" },
            "example": {
              "statusCode": 403,
              "message": "Insufficient role",
              "error": "Forbidden"
            }
          }
        }
      },
      "NotFound": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" },
            "example": {
              "statusCode": 404,
              "message": "Not found",
              "error": "Not Found"
            }
          }
        }
      }
    },
    "schemas": {
      "Role": {
        "type": "string",
        "enum": ["ADMIN", "CUSTOMER"]
      },
...
      "RegisterDto": {
        "type": "object",
        "required": ["fullName", "email", "password"],
        "properties": {
          "fullName": { "type": "string" },
          "email": { "type": "string", "format": "email" },
          "password": { "type": "string", "minLength": 6 },
          "phone": { "type": "string" }
        }
      },
      "LoginDto": {
        "type": "object",
        "required": ["email", "password"],
        "properties": {
          "email": { "type": "string", "format": "email" },
          "password": { "type": "string", "minLength": 6 }
        }
      },
...
      "CreateCategoryDto": {
        "type": "object",
        "required": ["name"],
        "properties": {
          "name": { "type": "string", "maxLength": 120 },
          "slug": { "type": "string", "maxLength": 160 },
          "description": { "type": "string", "maxLength": 500 }
        }
      },
      "UpdateCategoryDto": {
        "allOf": [{ "$ref": "#/components/schemas/CreateCategoryDto" }]
      },
      "AddCartItemDto": {
        "type": "object",
        "required": ["laptopId", "quantity"],
        "properties": {
          "laptopId": { "type": "string", "format": "uuid" },
          "quantity": { "type": "integer", "minimum": 1 }
        }
      },
      "UpdateCartItemDto": {
        "type": "object",
        "required": ["quantity"],
        "properties": {
          "quantity": { "type": "integer", "minimum": 1 }
        }
      },
...
      "UserPublic": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "fullName": { "type": "string" },
          "email": { "type": "string", "format": "email" },
          "phone": { "type": "string", "nullable": true },
          "role": { "$ref": "#/components/schemas/Role" },
          "isActive": { "type": "boolean" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
      "UserLoginView": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "fullName": { "type": "string" },
          "email": { "type": "string", "format": "email" },
          "role": { "$ref": "#/components/schemas/Role" }
        }
      },
      "LaptopImageKind": {
        "type": "string",
        "enum": ["FEATURED", "GALLERY"]
      },
      "LaptopImage": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "kind": { "$ref": "#/components/schemas/LaptopImageKind" },
          "url": { "type": "string", "format": "uri" },
          "publicId": { "type": "string" },
          "position": { "type": "integer" }
        }
      },
      "Laptop": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "title": { "type": "string" },
          "slug": { "type": "string" },
          "brand": { "type": "string" },
          "model": { "type": "string" },
          "price": { "type": "string", "example": "76999.00" },
          "stock": { "type": "integer" },
          "shortDescription": { "type": "string", "nullable": true },
          "description": { "type": "string", "nullable": true },
          "cpu": { "type": "string", "nullable": true },
          "ram": { "type": "string", "nullable": true },
          "storage": { "type": "string", "nullable": true },
          "gpu": { "type": "string", "nullable": true },
          "screenSize": { "type": "string", "nullable": true },
          "os": { "type": "string", "nullable": true },
          "categoryId": { "type": "string", "format": "uuid", "nullable": true },
          "category": {
            "oneOf": [
              { "$ref": "#/components/schemas/Category" },
              { "type": "null" }
            ]
          },
          "featuredImage": {
            "oneOf": [
              { "$ref": "#/components/schemas/LaptopImage" },
              { "type": "null" }
            ]
          },
          "galleryImages": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/LaptopImage" }
          },
          "isPublished": { "type": "boolean" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
      "Category": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "name": { "type": "string" },
          "slug": { "type": "string" },
          "description": { "type": "string", "nullable": true },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
      "CartItemWithLaptop": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "cartId": { "type": "string", "format": "uuid" },
          "laptopId": { "type": "string", "format": "uuid" },
          "quantity": { "type": "integer" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" },
          "laptop": { "$ref": "#/components/schemas/Laptop" }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "statusCode": { "type": "integer" },
          "message": {
            "oneOf": [
              { "type": "string" },
              {
                "type": "array",
                "items": { "type": "string" }
              }
            ]
          },
          "error": { "type": "string" }
        }
      },
...
    }
  },
...
}